WindowsVisualStudio2017patch
This is the windows visual studio patch. The release versions are stable and will run the game developement editor without a source editor. However the game's source code is developed natively in linux with clion. So this patch is required for the game editor to compile. This patch must downloaded and built. Afterwards it is possible to create a new branch and continue development from there. 

1. Download patch
2. compile engine/editor
3. Create new branch. 

It may be possible to use the Intermediate folder files from this patch to switch over to another previous branch, but I've yet to test it thoroughly. TODO